Use OlpClientSettings from the core.
Changes imports to use OlpClientSettings from the olp-sdk-core
in the olp-sdk-dataservice-read.

Also move FetchOptions to the core.

Updates for some tests.

Resolves: OLPEDGE-1940
