+++ @here/olp-sdk-dataservice-read/lib/client/QueryClient.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 HERE Europe B.V.
+ * Copyright (C) 2019-2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,22 +17,20 @@
  * License-Filename: LICENSE
  */
 
+import { FetchOptions, HRN, OlpClientSettings } from "@here/olp-sdk-core";
 import {
     AdditionalFields,
     MetadataApi,
     QueryApi
 } from "@here/olp-sdk-dataservice-api";
 import {
-    HRN,
     isValid,
     mortonCodeFromQuadKey,
-    OlpClientSettings,
     PartitionsRequest,
     QuadTreeIndexRequest,
     RequestFactory
 } from "..";
 import { MetadataCacheRepository } from "../cache/MetadataCacheRepository";
-import { FetchOptions } from "./FetchOptions";
 
 /**
  * A client for the platform Query Service that provides a way to get information (metadata)
