+++ @here/olp-sdk-dataservice-read/lib/client/SchemaDetailsRequest.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 HERE Europe B.V.
+ * Copyright (C) 2019-2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,8 @@
  * License-Filename: LICENSE
  */
 
-import { HRN, validateBillingTag } from "..";
+import { HRN } from "@here/olp-sdk-core";
+import { validateBillingTag } from "..";
 
 /**
  * Prepares information for calls to the platform Artifact Service.
