+++ @here/olp-sdk-dataservice-read/test/unit/StreamLayerClient.test.ts
@@ -101,7 +101,7 @@ describe("StreamLayerClient", () => {
         getBlobStub = sandbox.stub(BlobApi, "getBlob");
         subscribeStub = sandbox.stub(StreamApi, "subscribe");
         pollStub = sandbox.stub(StreamApi, "consumeData");
-        commitOffsetsStub = sandbox.stub(StreamApi, "commitOffsets");
+        commitOffsetsStub = sandbox.stub(StreamApi, "doCommitOffsets");
         unsubscribeStub = sandbox.stub(StreamApi, "deleteSubscription");
         seekOffsetsStub = sandbox.stub(StreamApi, "seekToOffset");
         getBaseUrlRequestStub = sandbox.stub(
